requirements.txt
beautifulSoup4
requests
langchain
langchain-google-genai
python-dotenv
langchain-mongodb
pymongo
streamlit
